Add function createProductImage
- Added service function createProductImage which creates an image for product/product variant in
the database and on the server itself. The purpose of this function is so that it can be used to create images
for products (which do not require image on Stripe) and product variants (which do require image on Stripe).
This way we do not have to use the exact same logic in multiple locations we can just use this function.

- Replaced the logic in the productImageController to use this service function when creating images