Add function createDatabaseAndStripeProduct and cloneProductImages
- Added service function createDatabaseAndStripeProductImage which creates an image for a product variant in the
database, on the server itself, and on Stripe. Function is to be used on product variants as product variant
images are stored in the database to be displayed in shop, and also stored in Stripe server so its displayed at
checkout.

- Added function cloneProductImages which clones the images belonging to a source product variant and adds them
to another product variant. It adds the images to the database and Stripe. Purpose of this function is so in
the admin dashboard, new product variants that are added to a product with a different size but same everything
else as other product variants, there is no need to manually re add all of the same images. You can clone the
images of a product variant that already exists and has the correct images.