Move functions from product variant service to product service
- Moved the createDatabaseProductImage and deleteDatabaseProductImage to productService as those functions will be called
directly in the productController. So it makes more sense to have them in their own service for organization.

- Modified the the function which deletes product images and all product variant images to also delete the product
and product variants. Renamed the function to deleteProduct. Did this as this function would only be called when
deleting the product so just added the functionality to delete the product and the images.