Modify function createDatabaseProductImage
- Changed name of function createDatabaseProductImage to createProductImage. Did this as the function does not only create
the image in the database but also on the server.

- Modified function createDatabaseProductImage so that if no buffer is given, image will only be added to the database not
on the server. Did this so that product variants which have the same image, there will be no need to upload the same image
to the server but for different product variants. This way the function is more dynamic so that product variants will be
able to share images on the server.