Modify controller function updateProductVariant
- Renamed controller function updateProductVariant to updateProductVariantQuantity. Did this as function now can only
update the quantity of a product variant. The reasoning for this is that price will always be updated for an entire group
and color will never be changed as groups are determined by product id and color. So if the color of a product variant
changes it will be apart of another group but still have the images of the previous group and everything will be ruined.