Add client services function adminProductsService
- Added client services function adminProductsService which sends a request to the server to perform CRUD product
operations. Created this function instead of having to create a seperate function for each CRUD operation.

- Modified controller function updateProduct to not be able to update the product name. Did this as the product variants
of product have their name on Stripe based on the products name. So if the products name is changed, you would have to
go through every products product variant on Stripe and also change their names. Do not have enough time for this.