Modify controller function updateProduct
- Modified product controller function updateProduct so it no longer just updates a products description. The
function now updates a products description/price in the database and all product variant prices belonging to
product in the database and on Stripe. Did this with user experience in mind when it comes to the admin
dashboard. If admin wants to update price for all product variants belonging to product they would have
to do so one product variant group at a time. Also, in reality all products of the same type no matter color
and size are the same price.