Remove ImageForm from ProductCard component
- Removed the image form from the product card component as it is wasteful to have an image form render for every
single product card. Instead the image form has been added to the products table and will be available for use by
all product cards.

- Removed the add image function from the products table and added it to the image form

- Added function updateSelectedProduct to the products table which updates the currently selected product. This
way in the image form, after adding an image the image form will get the updated product with the new image added.
No need to close the form and reopen it.